server/requirements.txt
Django==2.1.4
djangorestframework==3.9.0
graphene-django==2.2.0
django-filter==2.0.0
django-cors-headers==2.4.0
flake8==3.6.0
gunicorn==20.0.4